Make taskCallStringSize configurable
taskCallStringSize determines the size of worker query string.
It was originally hard coded to a specific value. This has caused
issues at some users. Since it determines initial shared memory
allocation, we did not want to set it to an arbitrary higher number.
Instead made it configurable.

This commit introduces a new GUC variable
max_worker_task_call_string_size..

Changes in this variable requires restart to be in effect.